moved RedesignedArbovirusPageSectionId to a constants file
